Assertion -> parent class
