+++ app/jobs/ledger/assert_no_orphaned_cts_job.rb
@@ -1,11 +1,8 @@
 # frozen_string_literal: true
 
 class Ledger
-  class AssertNoOrphanedCtsJob < ApplicationJob
-    queue_as :low
-    include AssertsRequirements
-
-    def run
+  class AssertNoOrphanedCtsJob < AssertRequirementJob
+    def perform
       @cts = CanonicalTransaction.where(ledger_item_id: nil)
 
       @cts.find_each do |ct|
